Fix datapoint row duplication bug
This was a really dumb bug. See https://hexdocs.pm/phoenix_live_view/form-bindings.html#pitfalls-with-input-names
Because names of form elements shadow the built-in properties of the form element(!), a form input with a name of "id" causes all sorts of trouble. The solution is to name it anything else, in this case "datapoint-id".